Init memory nodes on startup
