+++ internal/router/routes_test.go
@@ -2,11 +2,20 @@ package router
 
 import (
 	"fmt"
-	"github.com/stretchr/testify/assert"
 	"testing"
+
+	"github.com/NodeFactoryIo/vedran/internal/controllers"
+	mocks "github.com/NodeFactoryIo/vedran/mocks/models"
+	"github.com/gorilla/mux"
+	"github.com/stretchr/testify/assert"
 )
 
 func TestApiRouter(t *testing.T) {
+	nodeRepoMock := mocks.NodeRepository{}
+	pingRepoMock := mocks.PingRepository{}
+	metricsRepoMock := mocks.MetricsRepository{}
+	apiController := controllers.NewApiController(false, &nodeRepoMock, &pingRepoMock, &metricsRepoMock)
+
 	tests := []struct {
 		name    string
 		url     string
@@ -16,8 +25,10 @@ func TestApiRouter(t *testing.T) {
 		{name: "Test ping route", url: "/api/v1/nodes/pings", methods: []string{"POST"}},
 		{name: "Test metrics route", url: "/api/v1/nodes/metrics", methods: []string{"PUT"}},
 	}
-	// pass nil as db instance as only routes are tested
-	router := CreateNewApiRouter(nil, false)
+
+	router := mux.NewRouter()
+	createRoutes(apiController, router)
+
 	for _, test := range tests {
 		t.Run(fmt.Sprintf("%s::%s", test.name, test.url), func(t *testing.T) {
 			rout := router.GetRoute(test.url)
